+++ koop_form/apps/report/tests/test_views.py
@@ -21,8 +21,72 @@ from factories.model_factories import (
     OrderFactory,
     ProfileFactory,
 )
+import pandas as pd
+from io import StringIO
 
 logger = logging.getLogger("django.server")
+
+
+
+@pytest.mark.django_db
+class MassOrderBoxReportDownloadView(TestCase):
+    def setUp(self):
+        self.url = reverse(
+            "mass-order-box-report-download"
+        )
+        self.user_1 = UserFactory(is_staff=True)
+        self.user_2 = UserFactory(is_staff=True)
+        self.client.force_login(self.user_1)
+        self.client.force_login(self.user_2)
+        self.producer_1 = ProducerFactory(name="yoyoma",short="YOYO")
+        self.producer_2 = ProducerFactory(name="elomaaa",short="EMAA")
+        self.product_1 = ProductFactory(producer=self.producer_1, price=5.5, name="warzywo")
+        self.product_2 = ProductFactory(producer=self.producer_2, price=14, name="cebula")
+        self.product_3 = ProductFactory(producer=self.producer_2, price=21, name="ziemniak")
+        self.order_1 = OrderFactory(order_number=1, user = self.user_1)
+        self.order_2 = OrderFactory(order_number=2, user=self.user_2)
+
+    def test_response_and_content(self):
+        #given
+        OrderItemFactory(order=self.order_1, product=self.product_1, quantity=Decimal(2.5))
+        OrderItemFactory(order=self.order_2, product=self.product_1, quantity=Decimal(2))
+        OrderItemFactory(order=self.order_1, product=self.product_2, quantity=Decimal(5))
+        OrderItemFactory(order=self.order_2, product=self.product_2, quantity=Decimal(0.5))
+        OrderItemFactory(order=self.order_1, product=self.product_3, quantity=Decimal(10))
+        OrderItemFactory(order=self.order_2, product=self.product_3, quantity=Decimal(20))
+        #when
+        response = self.client.get(self.url)
+        #then
+        df = pd.read_csv(StringIO(response.content.decode("utf-8")), sep=",", header=None)
+        self.assertEquals(response.status_code, 200)
+        self.assertTrue(((df[1] == "cebula" ) & (df[2] == "5" ) ).any())
+        self.assertTrue(((df[1] == "warzywo") & (df[2] == "2,5")).any())
+        self.assertTrue(((df[1] == "ziemniak") & (df[2] == "10")).any())
+        self.assertTrue(((df[5] == "cebula") & (df[6] == "0,5")).any())
+        self.assertTrue(((df[5] == "warzywo") & (df[6] == "2")).any())
+        self.assertTrue(((df[5] == "ziemniak") & (df[6] == "20")).any())
+
+    def test_duplicated_item(self):
+        # given
+        OrderItemFactory(order=self.order_1, product=self.product_1, quantity=Decimal(2.5))
+        OrderItemFactory(order=self.order_2, product=self.product_1, quantity=Decimal(2))
+        OrderItemFactory(order=self.order_1, product=self.product_2, quantity=Decimal(5))
+        OrderItemFactory(order=self.order_2, product=self.product_2, quantity=Decimal(0.5))
+        OrderItemFactory(order=self.order_2, product=self.product_2, quantity=Decimal(3.5)) # duplicated cebula
+        OrderItemFactory(order=self.order_1, product=self.product_3, quantity=Decimal(10))
+        OrderItemFactory(order=self.order_2, product=self.product_3, quantity=Decimal(20))
+        # when
+        response = self.client.get(self.url)
+        # then
+        df = pd.read_csv(StringIO(response.content.decode("utf-8")), sep=",", header=None)
+        self.assertEquals(response.status_code, 200)
+        self.assertTrue(((df[1] == "cebula") & (df[2] == "5")).any())
+        self.assertTrue(((df[1] == "warzywo") & (df[2] == "2,5")).any())
+        self.assertTrue(((df[1] == "ziemniak") & (df[2] == "10")).any())
+        self.assertTrue(((df[5] == "cebula") & (df[6] == "0,5")).any())
+        self.assertTrue(((df[5] == "cebula") & (df[6] == "3,5")).any())
+        self.assertTrue(((df[5] == "warzywo") & (df[6] == "2")).any())
+        self.assertTrue(((df[5] == "ziemniak") & (df[6] == "20")).any())
 
 
 @pytest.mark.django_db
@@ -32,7 +96,7 @@ class TestProducerProductsReportView(TestCase):
         self.client.force_login(self.user)
         self.producer = ProducerFactory()
         self.url = reverse(
-            "producer-products-report", kwargs={"slug": self.producer.slug}
+            "producer-orders-report", kwargs={"slug": self.producer.slug}
         )
         self.product1 = ProductFactory(
             producer=self.producer, price=5.5, name="warzywo"
@@ -59,9 +123,9 @@ class TestProducerProductsReportView(TestCase):
         assert context_data["producer"] == self.producer
         assert context_data["producers"] == get_producers_list(Producer)
         assert sorted(context_data["product_names_list"]) == ["cebula", "warzywo"]
-        assert sorted(context_data["product_ordered_quantities_list"]) == [4.5, 5.5]
-        assert sorted(context_data["product_incomes_list"]) == ["24.75", "77.00"]
-        assert context_data["total_income"] == 101.75
+        assert sorted(context_data["order_quantities_list"]) == [4.5, 5.5]
+        assert sorted(context_data["order_incomes_list"]) == ["24.75", "77.00"]
+        assert context_data["total_order_income"] == 101.75
 
     def test_user_is_not_staff(self):
         self.client.force_login(UserFactory())
@@ -101,11 +165,6 @@ class TestProducerBoxReportView(TestCase):
         assert context_data["producer"] == self.producer
         assert list(context_data["producers"]) == list(producers)
         assert list(context_data["products"]) == list(products)
-        assert list(context_data["order_data"]) == [
-            "(skrz1: 4) ",
-            "(skrz2: 1) ",
-            "(skrz3: 2) ",
-        ]
         assert response.status_code == 200
 
     def test_user_is_not_staff(self):
@@ -136,7 +195,6 @@ class TestUsersReportView(TestCase):
 
         assert response.status_code == 200
         assert context_data["user_name_list"] == ["K Kamil", "M Marek"]
-        assert context_data["user_order_number_list"] == [1, 2]
         assert context_data["user_pickup_day_list"] == ["środa", "czwartek"]
         assert context_data["user_phone_number_list"] == [444555666, 777666555]
 
@@ -157,7 +215,7 @@ class TestProducerProductsListView(TestCase):
             5,
         ):
             ProducerFactory()
-        self.url = reverse("producer-products-list")
+        self.url = reverse("producer-orders-list")
 
     def test_user_is_not_staff(self):
         response = self.client.get(self.url)