+++ koop_form/factories/model_factories.py
@@ -6,7 +6,6 @@ from factory.django import DjangoModelFactory
 import pytz
 from apps.form.models import Producer, Product, Status, WeightScheme, Order, OrderItem
 from django.utils import timezone
-
 from apps.user.models import UserProfile
 
 ModelUser = get_user_model()
@@ -17,6 +16,7 @@ class ProducerFactory(DjangoModelFactory):
         model = Producer
 
     name = Faker("name")
+    short = Faker("word")
     description = Faker("paragraph")
     order = Faker("random_int", min=1, max=10000)
 
@@ -65,6 +65,7 @@ class OrderFactory(DjangoModelFactory):
     pick_up_day = ["środa", "czwartek"][random.randint(0, 1)]
     date_created = timezone.now()
     is_given = False
+    order_number = 1
 
     @post_generation
     def products(self, create, extracted, **kwargs):