+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.0.2]
+### Fix
+- [marcinkosztolowicz] Fix error exporting all orders csv when item is duplicated in any order
+- [marcinkosztolowicz] Change MassOrderBoxReport items in box order by producer short name 
+- [marcinkosztolowicz] Change  items in producer product order by  box number
+
+# [1.0.1]
+### Fix
+- [marcinkosztolowicz] Fix http 500 when weight scheme 0.000 is added during product creation.
+
+# [1.0.0]
+### Added
+- [marcinkosztolowicz] Deployment IaC
+- [arturraczka] Code
 
 ## [0.1.0] - WIP
 ### Added